+++ Tests/MapboxMapsTests/Gestures/GestureHandlers/Mocks/MockRotationGestureRecognizer.swift
@@ -12,15 +12,31 @@ final class MockRotationGestureRecognizer: UIRotationGestureRecognizer {
         }
     }
 
-    let getRotationStub = Stub<Void, CGFloat>(defaultReturnValue: 2)
-    override var rotation: CGFloat {
+    let getVelocityStub = Stub<Void, CGFloat>(defaultReturnValue: 0)
+    override var velocity: CGFloat {
         get {
-            getRotationStub.call()
+            getVelocityStub.call()
         }
         // swiftlint:disable:next unused_setter_value
         set {
             fatalError("unimplemented")
         }
+    }
+
+    let getRotationStub = Stub<Void, CGFloat>(defaultReturnValue: 0)
+    let setRotationStub = Stub<CGFloat, Void>()
+    override var rotation: CGFloat {
+        get {
+            getRotationStub.call()
+        }
+        set {
+            setRotationStub.call(with: newValue)
+        }
+    }
+
+    let locationInViewStub = Stub<UIView?, CGPoint>(defaultReturnValue: .zero)
+    override func location(in view: UIView?) -> CGPoint {
+        locationInViewStub.call(with: view)
     }
 
     struct AddTargetParams {